AI's no longer have active abilities outside of their basic attack
